move controllers into platform namespace
