+++ routes/web.php
@@ -52,14 +52,6 @@ Route::middleware('auth')->namespace('Admin')->prefix('settings')->group(functio
     Route::get('permissions', 'PermissionsController@index');
     Route::post('permissions', 'PermissionsController@update');
 
-    Route::get('keys', 'KeysController@index');
-    Route::post('keys', 'KeysController@store');
-    Route::delete('keys/{key}', 'KeysController@destroy');
-
-    Route::get('clients', 'ClientsController@index');
-    Route::post('clients', 'ClientsController@store');
-    Route::delete('clients/{client}', 'ClientsController@destroy');
-
     Route::get('users', 'UsersController@index');
     Route::post('users', 'UsersController@store');
     Route::post('users/{user}', 'UsersController@update');
